pyproject.toml
[tool.poetry]
name = "test-rates"
version = "0.1.0"
description = ""
authors = ["Vladislav-76 <sillmaril2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
